# ♻️ Cleanup and optimize `MerkleTrie.Get()`